backend/requirements.txt
# Django & Web Framework
Django==5.1
djangorestframework==3.14.0
gunicorn==21.2.0
whitenoise==6.6.0
drf-spectacular==0.27.0

# Database
psycopg[binary]==3.2.13
dj-database-url==2.1.0
python-decouple==3.8

# Security
cryptography>=41.0.0

# PWA / Push Notifications
pywebpush>=1.14.0
